requirements.txt
telethon==1.42.0
httpx==0.28.1
python-dotenv==1.0.1
aiosqlite==0.20.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.12
